Add and use getCurrentDocContext
